products/isb1/pyproject.toml
[project]
name = "inferscope-benchmark"
version = "1.0.0"
description = "ISB-1: Inference Serving Benchmark Standard 1 for reproducible LLM serving evaluation"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
    "requests>=2.31",
    "aiohttp>=3.9",
    "tiktoken>=0.5",
    "rouge-score>=0.1.2",
    "tqdm>=4.65",
    "rich>=13.0",
    "click>=8.1",
    "jinja2>=3.1",
    "httpx>=0.25",
]

[project.optional-dependencies]
datasets = [
    "datasets>=2.14.0",
    "huggingface-hub>=0.20.0",
]
quality = [
    "human-eval",
    "lm-eval>=0.4",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
isb1 = "harness.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["harness*", "workloads*", "analysis*", "quality*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
